Refactored results objects to hold single values
