+++ Environment_oM/Environment_oM.csproj
@@ -64,36 +64,37 @@
     <Compile Include="Elements\Space.cs" />
     <Compile Include="Fragments\AnalyticalConstruction.cs" />
     <Compile Include="Fragments\LightReflectanceFragment.cs" />
-    <Compile Include="Results\IAnalysisResult.cs" />
-    <Compile Include="Results\ResultObjects\DiffuseHorizontalRadiation.cs" />
-    <Compile Include="Results\ResultObjects\DirectNormalRadiationResult.cs" />
-    <Compile Include="Results\ResultObjects\HumidityRatioResult.cs" />
-    <Compile Include="Results\ResultObjects\PowerResult.cs" />
-    <Compile Include="Results\ResultObjects\IEnergyPerUnitTimeResult.cs" />
-    <Compile Include="Results\ResultObjects\IEnergyResult.cs" />
-    <Compile Include="Results\ResultObjects\InsolationResult.cs" />
-    <Compile Include="Results\ResultObjects\GlobalHorizontalRadiationResult.cs" />
-    <Compile Include="Results\ResultObjects\IEnergyPerUnitAreaResult.cs" />
-    <Compile Include="Results\ResultObjects\DryBulbTemperatureResult.cs" />
-    <Compile Include="Results\ResultObjects\IlluminanceResult.cs" />
-    <Compile Include="Results\ResultObjects\UniversalThermalClimateIndexTemperatureResult.cs" />
-    <Compile Include="Results\ResultObjects\IRatioResult.cs" />
-    <Compile Include="Results\ResultObjects\SkyTemperatureResult.cs" />
-    <Compile Include="Results\ResultObjects\IAngleResult.cs" />
-    <Compile Include="Results\ResultObjects\DewPointTemperatureResult.cs" />
-    <Compile Include="Results\ResultObjects\SpecificEnergyResult.cs" />
-    <Compile Include="Results\ResultObjects\EnthalpyResult.cs" />
-    <Compile Include="Results\ResultObjects\IPressureResult.cs" />
-    <Compile Include="Results\ResultObjects\WindDirectionResult.cs" />
-    <Compile Include="Results\ResultObjects\AtmosphericPressureResult.cs" />
-    <Compile Include="Results\ResultObjects\WindSpeedResult.cs" />
-    <Compile Include="Results\ResultObjects\SurfaceTemperatureResult.cs" />
-    <Compile Include="Results\ResultObjects\IVelocityResult.cs" />
-    <Compile Include="Results\ResultObjects\WetBulbTemperatureResult.cs" />
-    <Compile Include="Results\ResultObjects\MeanRadiantTemperatureResult.cs" />
-    <Compile Include="Results\ResultObjects\RelativeHumidityResult.cs" />
-    <Compile Include="Results\ResultObjects\ITemperatureResult.cs" />
-    <Compile Include="Results\TimeIndexResult.cs" />
+    <Compile Include="Results\ResultObjects\IAnalysisResult.cs" />
+    <Compile Include="Results\ResultObjects\EnergyPerUnitArea\DiffuseHorizontalRadiation.cs" />
+    <Compile Include="Results\ResultObjects\EnergyPerUnitArea\DirectNormalRadiationResult.cs" />
+    <Compile Include="Results\ResultObjects\Ratio\HumidityRatioResult.cs" />
+    <Compile Include="Results\ResultObjects\EnergyPerUnitTime\PowerResult.cs" />
+    <Compile Include="Results\ResultObjects\EnergyPerUnitTime\IEnergyPerUnitTimeResult.cs" />
+    <Compile Include="Results\ResultObjects\Energy\IEnergyResult.cs" />
+    <Compile Include="Results\ResultObjects\EnergyPerUnitArea\InsolationResult.cs" />
+    <Compile Include="Results\ResultObjects\EnergyPerUnitArea\GlobalHorizontalRadiationResult.cs" />
+    <Compile Include="Results\ResultObjects\EnergyPerUnitArea\IEnergyPerUnitAreaResult.cs" />
+    <Compile Include="Results\ResultObjects\Temperature\DryBulbTemperatureResult.cs" />
+    <Compile Include="Results\ResultObjects\Illuminance\IlluminanceResult.cs" />
+    <Compile Include="Results\ResultObjects\Temperature\UniversalThermalClimateIndexTemperatureResult.cs" />
+    <Compile Include="Results\ResultObjects\Ratio\IRatioResult.cs" />
+    <Compile Include="Results\ResultObjects\Temperature\SkyTemperatureResult.cs" />
+    <Compile Include="Results\ResultObjects\Angle\IAngleResult.cs" />
+    <Compile Include="Results\ResultObjects\Temperature\DewPointTemperatureResult.cs" />
+    <Compile Include="Results\ResultObjects\SpecificEnergy\ISpecificEnergyResult.cs" />
+    <Compile Include="Results\ResultObjects\SpecificEnergy\EnthalpyResult.cs" />
+    <Compile Include="Results\ResultObjects\Pressure\IPressureResult.cs" />
+    <Compile Include="Results\ResultObjects\Angle\WindDirectionResult.cs" />
+    <Compile Include="Results\ResultObjects\Pressure\AtmosphericPressureResult.cs" />
+    <Compile Include="Results\ResultObjects\DateTimeIndex.cs" />
+    <Compile Include="Results\ResultObjects\Velocity\WindSpeedResult.cs" />
+    <Compile Include="Results\ResultObjects\Temperature\SurfaceTemperatureResult.cs" />
+    <Compile Include="Results\ResultObjects\Velocity\IVelocityResult.cs" />
+    <Compile Include="Results\ResultObjects\Temperature\WetBulbTemperatureResult.cs" />
+    <Compile Include="Results\ResultObjects\Temperature\MeanRadiantTemperatureResult.cs" />
+    <Compile Include="Results\ResultObjects\Ratio\RelativeHumidityResult.cs" />
+    <Compile Include="Results\ResultObjects\Temperature\ITemperatureResult.cs" />
+    <Compile Include="Results\ResultObjects\TimeIndexResult.cs" />
     <Compile Include="SpaceCriteria\Acoustics.cs" />
     <Compile Include="SpaceCriteria\Enums\EmitterType.cs" />
     <Compile Include="SpaceCriteria\Enums\ProfileDay.cs" />
