+++ Environment_oM/Results/ResultObjects/Angle/IAngleResult.cs
