+++ Environment_oM/Results/ResultObjects/Angle/WindDirectionResult.cs
@@ -20,23 +20,24 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
-using BH.oM.Base;
-using BH.oM.Quantities.Attributes;
 using System.Collections.Generic;
 using System.ComponentModel;
 
+using BH.oM.Base;
+using BH.oM.Quantities.Attributes;
+
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing energy-flux data")]
-    public class DirectNormalRadiationResult : BHoMObject, IEnergyPerUnitAreaResult
+    [Description("A results object containing wind-direction data")]
+    public class WindDirectionResult : BHoMObject, IAngleResult
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        [EnergyPerUnitArea]
+        [Angle]
         [Description("The value associated with the results object")]
-        public virtual List<double> Result { get; set; } = new List<double>();
+        public virtual double Result { get; set; } = new double();
 
         /***************************************************/
     }
