+++ Environment_oM/Results/ResultObjects/Energy/IEnergyResult.cs
@@ -20,25 +20,16 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
+using BH.oM.Base;
+using BH.oM.Quantities.Attributes;
 using System.Collections.Generic;
 using System.ComponentModel;
 
-using BH.oM.Base;
-using BH.oM.Quantities.Attributes;
-
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing temperature data")]
-    public class SurfaceTemperatureResult : BHoMObject, ITemperatureResult
+    [Description("A results object containing energy data")]
+    public interface IEnergyResult : IBHoMObject, IAnalysisResult
     {
-        /***************************************************/
-        /**** Properties                                ****/
-        /***************************************************/
 
-        [Temperature]
-        [Description("The value associated with the results object")]
-        public virtual List<double> Result { get; set; } = new List<double>();
-
-        /***************************************************/
     }
 }
