+++ Environment_oM/Results/ResultObjects/EnergyPerUnitArea/DiffuseHorizontalRadiation.cs
@@ -27,9 +27,17 @@ using System.ComponentModel;
 
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing energy/power data")]
-    public interface IEnergyResult : IBHoMObject, IAnalysisResult
+    [Description("A results object containing diffuse horizontal radiation data")]
+    public class DiffuseHorizontalRadiationResult : BHoMObject, IEnergyPerUnitAreaResult
     {
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
 
+        [EnergyPerUnitArea]
+        [Description("The value associated with the results object")]
+        public virtual double Result { get; set; } = new double();
+
+        /***************************************************/
     }
 }
