+++ Environment_oM/Results/ResultObjects/EnergyPerUnitArea/GlobalHorizontalRadiationResult.cs
@@ -21,23 +21,25 @@
  */
 
 using BH.oM.Base;
+using BH.oM.Quantities.Attributes;
 using System.Collections.Generic;
 using System.ComponentModel;
-using BH.oM.Quantities.Attributes;
 
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing relative humidity data")]
-    public class RelativeHumidityResult : BHoMObject, IRatioResult
+    [Description("A results object containing global horizontal radiation data")]
+    public class GlobalHorizontalRadiationResult : BHoMObject, IEnergyPerUnitAreaResult
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        [Ratio]
+        [EnergyPerUnitArea]
         [Description("The value associated with the results object")]
-        public virtual List<double> Result { get; set; } = new List<double>();
+        public virtual double Result { get; set; } = new double();
 
         /***************************************************/
     }
 }
+
+
