+++ Environment_oM/Results/ResultObjects/EnergyPerUnitTime/PowerResult.cs
@@ -21,13 +21,23 @@
  */
 
 using BH.oM.Base;
+using System.Collections.Generic;
 using System.ComponentModel;
+using BH.oM.Quantities.Attributes;
 
 namespace BH.oM.Environment.Results
 {
     [Description("A results object containing power data")]
-    public interface IEnergyPerUnitTimeResult : IBHoMObject, IAnalysisResult
+    public class PowerResult : BHoMObject, IEnergyPerUnitTimeResult
     {
+        /***************************************************/
+        /**** Properties                                ****/
+        /***************************************************/
 
+        [EnergyPerUnitTime]
+        [Description("The value associated with the results object")]
+        public virtual double Result { get; set; } = new double();
+
+        /***************************************************/
     }
 }
