+++ Environment_oM/Results/ResultObjects/Illuminance/IlluminanceResult.cs
@@ -27,19 +27,17 @@ using BH.oM.Quantities.Attributes;
 
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing relative humidity data")]
-    public class HumidityRatioResult : BHoMObject, IRatioResult
+    [Description("A results object containing illuminance data")]
+    public class IlluminanceResult : BHoMObject, IAnalysisResult
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        [Ratio]
+        [Quantities.Attributes.Illuminance] // TODO - for some reason not including traceback to Quantities results in error here
         [Description("The value associated with the results object")]
-        public virtual List<double> Result { get; set; } = new List<double>();
+        public virtual double Result { get; set; } = new double();
 
         /***************************************************/
     }
 }
-
-
