+++ Environment_oM/Results/ResultObjects/Ratio/IRatioResult.cs
