+++ Environment_oM/Results/ResultObjects/Ratio/RelativeHumidityResult.cs
@@ -20,24 +20,23 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
+using BH.oM.Base;
 using System.Collections.Generic;
 using System.ComponentModel;
-
-using BH.oM.Base;
 using BH.oM.Quantities.Attributes;
 
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing temperature data")]
-    public class SkyTemperatureResult : BHoMObject, ITemperatureResult
+    [Description("A results object containing relative humidity data")]
+    public class RelativeHumidityResult : BHoMObject, IRatioResult
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        [Temperature]
+        [Ratio]
         [Description("The value associated with the results object")]
-        public virtual List<double> Result { get; set; } = new List<double>();
+        public virtual double Result { get; set; } = new double();
 
         /***************************************************/
     }
