+++ Environment_oM/Results/ResultObjects/SpecificEnergy/ISpecificEnergyResult.cs
@@ -21,24 +21,13 @@
  */
 
 using BH.oM.Base;
-using System.Collections.Generic;
 using System.ComponentModel;
-
-using BH.oM.Quantities.Attributes;
 
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing radiant temperature data")]
-    public class MeanTemperatureResult : BHoMObject, ITemperatureResult
+    [Description("A results object containing specific energy data")]
+    public interface ISpecificEnergyResult : IBHoMObject, IAnalysisResult
     {
-        /***************************************************/
-        /**** Properties                                ****/
-        /***************************************************/
 
-        [Temperature]
-        [Description("The value associated with the results object")]
-        public virtual List<double> Result { get; set; } = new List<double>();
-
-        /***************************************************/
     }
 }
