+++ Environment_oM/Results/ResultObjects/Temperature/ITemperatureResult.cs
