+++ Environment_oM/Results/ResultObjects/Temperature/MeanRadiantTemperatureResult.cs
@@ -21,22 +21,23 @@
  */
 
 using BH.oM.Base;
-using BH.oM.Quantities.Attributes;
 using System.Collections.Generic;
 using System.ComponentModel;
 
+using BH.oM.Quantities.Attributes;
+
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing atmospheric pressure data")]
-    public class AtmosphericPressureResult : BHoMObject, IPressureResult
+    [Description("A results object containing mean radiant radiant temperature data")]
+    public class MeanTemperatureResult : BHoMObject, ITemperatureResult
     {
         /***************************************************/
         /**** Properties                                ****/
         /***************************************************/
 
-        [Pressure]
+        [Temperature]
         [Description("The value associated with the results object")]
-        public virtual List<double> Result { get; set; } = new List<double>();
+        public virtual double Result { get; set; } = new double();
 
         /***************************************************/
     }
