+++ Environment_oM/Results/ResultObjects/TimeIndexResult.cs
@@ -1,4 +1,4 @@
-/*
+﻿/*
  * This file is part of the Buildings and Habitats object Model (BHoM)
  * Copyright (c) 2015 - 2021, the respective contributors. All rights reserved.
  *
@@ -20,14 +20,16 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
+using System.Collections.Generic;
+
 using BH.oM.Base;
-using System.ComponentModel;
 
 namespace BH.oM.Environment.Results
 {
-    [Description("A results object containing energy-flux data")]
-    public interface IEnergyPerUnitAreaResult : IBHoMObject, IAnalysisResult
+    public class TimeIndexResult : BHoMObject
     {
+        public virtual List<IAnalysisResult> Results { get; set; } = new List<IAnalysisResult>();
 
+        public virtual int TimeIndex { get; set; } = new int();
     }
 }
