+++ Environment_oM/Results/ResultObjects/Velocity/IVelocityResult.cs
